// frontend/src/types/AppointmentReportView.tsx
type AppointmentReportView = {
  appointmentId: number
  title: string
  description: string
  stylistId: number
  stylist: string
  stylistEmail: string
  stylistPhone: string
  client: string
  clientEmail: string
  clientPhone: string
  startTime: string
  endTime: string
  userId: number
  username: string
}

export default AppointmentReportView